Break new NUX view controller file into logical parts, and reorg NUX project folders
