+++ WordPress/Classes/ViewRelated/NUX/LoginViewController.swift
@@ -1,0 +1,180 @@
+/// View Controller for login-specific screens
+class LoginViewController: NUXViewController, SigninWPComSyncHandler, LoginFacadeDelegate {
+    @IBOutlet var instructionLabel: UILabel?
+    @objc var errorToPresent: Error?
+    var restrictToWPCom = false
+
+    lazy var loginFacade: LoginFacade = {
+        let facade = LoginFacade()
+        facade.delegate = self
+        return facade
+    }()
+
+    // MARK: Lifecycle Methods
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        displayError(message: "")
+        setupNavBarIcon()
+        styleInstructions()
+
+        if let error = errorToPresent {
+            displayRemoteError(error)
+        }
+    }
+
+
+    // MARK: - Setup and Configuration
+
+    /// Places the WordPress logo in the navbar
+    ///
+    @objc func setupNavBarIcon() {
+        addWordPressLogoToNavController()
+    }
+
+    /// Configures instruction label font
+    ///
+    @objc func styleInstructions() {
+        instructionLabel?.font = WPStyleGuide.mediumWeightFont(forStyle: .subheadline)
+    }
+
+    func configureViewLoading(_ loading: Bool) {
+        configureSubmitButton(animating: loading)
+        navigationItem.hidesBackButton = loading
+    }
+
+    /// Sets the text of the error label.
+    func displayError(message: String) {
+        guard message.count > 0 else {
+            errorLabel?.isHidden = true
+            return
+        }
+        errorLabel?.isHidden = false
+        errorLabel?.text = message
+    }
+
+    fileprivate func shouldShowEpilogue() -> Bool {
+        if !isJetpackLogin() {
+            return true
+        }
+        let context = ContextManager.sharedInstance().mainContext
+        let accountService = AccountService(managedObjectContext: context)
+        guard
+            let objectID = loginFields.meta.jetpackBlogID,
+            let blog = context.object(with: objectID) as? Blog,
+            let account = blog.account
+            else {
+                return false
+        }
+        return accountService.isDefaultWordPressComAccount(account)
+    }
+
+    func dismiss() {
+        if shouldShowEpilogue() {
+            self.performSegue(withIdentifier: .showEpilogue, sender: self)
+            return
+        }
+        dismissBlock?(false)
+        navigationController?.dismiss(animated: true, completion: nil)
+    }
+
+    /// Validates what is entered in the various form fields and, if valid,
+    /// proceeds with login.
+    ///
+    func validateFormAndLogin() {
+        view.endEditing(true)
+        displayError(message: "")
+
+        // Is everything filled out?
+        if !SigninHelpers.validateFieldsPopulatedForSignin(loginFields) {
+            let errorMsg = NSLocalizedString("Please fill out all the fields", comment: "A short prompt asking the user to properly fill out all login fields.")
+            displayError(message: errorMsg)
+
+            return
+        }
+
+        configureViewLoading(true)
+
+        loginFacade.signIn(with: loginFields)
+    }
+
+    /// Manages data transfer when seguing to a new VC
+    ///
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        guard let source = segue.source as? LoginViewController else {
+            return
+        }
+
+        if let destination = segue.destination as? LoginEpilogueViewController {
+            destination.dismissBlock = source.dismissBlock
+            destination.jetpackLogin = source.loginFields.meta.jetpackLogin
+        } else if let destination = segue.destination as? LoginViewController {
+            destination.loginFields = source.loginFields
+            destination.restrictToWPCom = source.restrictToWPCom
+            destination.dismissBlock = source.dismissBlock
+            destination.errorToPresent = source.errorToPresent
+        }
+    }
+
+    // MARK: SigninWPComSyncHandler methods
+    dynamic func finishedLogin(withUsername username: String!, authToken: String!, requiredMultifactorCode: Bool) {
+        syncWPCom(username, authToken: authToken, requiredMultifactor: requiredMultifactorCode)
+        guard let service = loginFields.meta.socialService, service == SocialServiceName.google,
+            let token = loginFields.meta.socialServiceIDToken else {
+                return
+        }
+
+        let accountService = AccountService(managedObjectContext: ContextManager.sharedInstance().mainContext)
+        accountService.connectToSocialService(service, serviceIDToken: token, success: {
+            WPAppAnalytics.track(.loginSocialConnectSuccess)
+            WPAppAnalytics.track(.loginSocialSuccess)
+        }, failure: { error in
+            DDLogError(error.description)
+            WPAppAnalytics.track(.loginSocialConnectFailure, error: error)
+            // We're opting to let this call fail silently.
+            // Our user has already successfully authenticated and can use the app --
+            // connecting the social service isn't critical.  There's little to
+            // be gained by displaying an error that can not currently be resolved
+            // in the app and doing so might tarnish an otherwise satisfying login
+            // experience.
+            // If/when we add support for manually connecting/disconnecting services
+            // we can revisit.
+        })
+    }
+
+    func isJetpackLogin() -> Bool {
+        return loginFields.meta.jetpackLogin
+    }
+
+    func configureStatusLabel(_ message: String) {
+        // this is now a no-op, unless status labels return
+    }
+
+    /// Overridden here to direct these errors to the login screen's error label
+    dynamic func displayRemoteError(_ error: Error!) {
+        configureViewLoading(false)
+
+        let err = error as NSError
+        guard err.code != 403 else {
+            let message = NSLocalizedString("Whoops, something went wrong and we couldn't log you in. Please try again!", comment: "An error message shown when a wpcom user provides the wrong password.")
+            displayError(message: message)
+            return
+        }
+
+        displayError(err, sourceTag: sourceTag)
+    }
+
+    func needsMultifactorCode() {
+        displayError(message: "")
+        configureViewLoading(false)
+
+        WPAppAnalytics.track(.twoFactorCodeRequested)
+        self.performSegue(withIdentifier: .show2FA, sender: self)
+    }
+
+    // Update safari stored credentials. Call after a successful sign in.
+    ///
+    func updateSafariCredentialsIfNeeded() {
+        SigninHelpers.updateSafariCredentialsIfNeeded(loginFields)
+    }
+}
