+++ WordPress/Classes/ViewRelated/NUX/NUXViewControllerBase.swift
@@ -1,0 +1,189 @@
+import Gridicons
+
+/// base protocol for NUX view controllers
+protocol NUXViewControllerBase {
+    var sourceTag: SupportSourceTag { get }
+    var helpBadge: WPNUXHelpBadgeLabel { get }
+    var helpButton: UIButton { get }
+    var loginFields: LoginFields { get }
+    var dismissBlock: ((_ cancelled: Bool) -> Void)? { get }
+
+    /// Checks if the signin vc modal should show a back button. The back button
+    /// visible when there is more than one child vc presented, and there is not
+    /// a case where a `SigninChildViewController.backButtonEnabled` in the stack
+    /// returns false.
+    ///
+    /// - Returns: True if the back button should be visible. False otherwise.
+    ///
+    func shouldShowCancelButton() -> Bool
+    func setupCancelButtonIfNeeded()
+}
+
+/// extension for NUXViewControllerBase where the base class is UIViewController (and thus also NUXTableViewController)
+extension NUXViewControllerBase where Self: UIViewController, Self: UIViewControllerTransitioningDelegate {
+
+    func shouldShowCancelButtonBase() -> Bool {
+        return isCancellable() && navigationController?.viewControllers.first == self
+    }
+
+    /// Sets up the cancel button for the navbar if its needed.
+    /// The cancel button is only shown when its appropriate to dismiss the modal view controller.
+    ///
+    func setupCancelButtonIfNeeded() {
+        if !shouldShowCancelButton() {
+            return
+        }
+
+        let cancelButton = UIBarButtonItem(barButtonSystemItem: .cancel, target: nil, action: nil)
+        cancelButton.on() { [weak self] (control: UIBarButtonItem) in
+            self?.handleCancelButtonTapped()
+        }
+        navigationItem.leftBarButtonItem = cancelButton
+    }
+
+    /// Checks if the signin vc modal should be cancellable. The controller is
+    /// cancellable when there is a default wpcom account, or at least one
+    /// self-hosted blog.
+    ///
+    /// - Returns: True if cancellable. False otherwise.
+    ///
+    func isCancellable() -> Bool {
+        // if there is an existing blog, or an existing account return true.
+        let context = ContextManager.sharedInstance().mainContext
+        let blogService = BlogService(managedObjectContext: context)
+
+        return AccountHelper.isDotcomAvailable() || blogService.blogCountForAllAccounts() > 0
+    }
+
+    /// Displays a login error in an attractive dialog
+    ///
+    func displayError(_ error: NSError, sourceTag: SupportSourceTag) {
+        let presentingController = navigationController ?? self
+        let controller = FancyAlertViewController.alertForError(error as NSError, loginFields: loginFields, sourceTag: sourceTag)
+        controller.modalPresentationStyle = .custom
+        controller.transitioningDelegate = self
+        presentingController.present(controller, animated: true, completion: nil)
+    }
+
+    /// Displays a login error message in an attractive dialog
+    ///
+    func displayErrorAlert(_ message: String, sourceTag: SupportSourceTag) {
+        let presentingController = navigationController ?? self
+        let controller = FancyAlertViewController.alertForGenericErrorMessageWithHelpshiftButton(message, loginFields: loginFields, sourceTag: sourceTag)
+        controller.modalPresentationStyle = .custom
+        controller.transitioningDelegate = self
+        presentingController.present(controller, animated: true, completion: nil)
+    }
+
+    /// It is assumed that NUX view controllers are always presented modally.
+    ///
+    func dismiss() {
+        dismiss(cancelled: false)
+    }
+
+    /// It is assumed that NUX view controllers are always presented modally.
+    /// This method dismisses the view controller
+    ///
+    /// - Parameters:
+    ///     - cancelled: Should be passed true only when dismissed by a tap on the cancel button.
+    ///
+    fileprivate func dismiss(cancelled: Bool) {
+        dismissBlock?(cancelled)
+        self.dismiss(animated: true, completion: nil)
+    }
+
+    // MARK: - Notifications
+
+    /// Updates the badge count and its visibility.
+    ///
+    func handleHelpshiftUnreadCountUpdated(_ notification: Foundation.Notification) {
+        let count = HelpshiftUtils.unreadNotificationCount()
+        helpBadge.text = "\(count)"
+        helpBadge.isHidden = (count == 0)
+    }
+
+
+    // MARK: - Actions
+
+    func handleBackgroundTapGesture(_ tgr: UITapGestureRecognizer) {
+        view.endEditing(true)
+    }
+
+    func handleCancelButtonTapped() {
+        dismiss(cancelled: true)
+        NotificationCenter.default.post(name: .WPLoginCancelled, object: nil)
+    }
+
+    // Handle the help button being tapped
+    //
+    func handleHelpButtonTapped(_ sender: AnyObject) {
+        displaySupportViewController(sourceTag: sourceTag)
+    }
+
+
+    // MARK: - Navbar Help and WP Logo methods
+
+    /// Adds the WP logo to the nav controller
+    func addWordPressLogoToNavController() {
+        let image = Gridicon.iconOfType(.mySites)
+        let imageView = UIImageView(image: image.imageWithTintColor(UIColor.white))
+        navigationItem.titleView = imageView
+    }
+
+    func addHelpButtonToNavController() {
+        let helpButtonMarginSpacerWidth = CGFloat(-8)
+        let helpBadgeSize = CGSize(width: 12, height: 10)
+        let helpButtonContainerFrame = CGRect(x: 0, y: 0, width: 44, height: 44)
+
+        NotificationCenter.default.addObserver(forName: .HelpshiftUnreadCountUpdated, object: nil, queue: nil) { [weak self](notification) in
+            self?.handleHelpshiftUnreadCountUpdated(notification)
+        }
+
+        let customView = UIView(frame: helpButtonContainerFrame)
+
+        helpButton.setTitle(NSLocalizedString("Help", comment: "Help button"), for: .normal)
+        helpButton.setTitleColor(UIColor(white: 1.0, alpha: 0.4), for: .highlighted)
+        helpButton.on(.touchUpInside) { [weak self] control in
+            guard let strongSelf = self else {
+                return
+            }
+            strongSelf.handleHelpButtonTapped(strongSelf.helpButton)
+        }
+
+        customView.addSubview(helpButton)
+        helpButton.translatesAutoresizingMaskIntoConstraints = false
+        helpButton.leadingAnchor.constraint(equalTo: customView.leadingAnchor).isActive = true
+        helpButton.trailingAnchor.constraint(equalTo: customView.trailingAnchor).isActive = true
+        helpButton.topAnchor.constraint(equalTo: customView.topAnchor).isActive = true
+        helpButton.bottomAnchor.constraint(equalTo: customView.bottomAnchor).isActive = true
+
+        helpBadge.translatesAutoresizingMaskIntoConstraints = false
+        helpBadge.isHidden = true
+        customView.addSubview(helpBadge)
+        helpBadge.centerXAnchor.constraint(equalTo: helpButton.trailingAnchor).isActive = true
+        helpBadge.centerYAnchor.constraint(equalTo: helpButton.topAnchor).isActive = true
+        helpBadge.widthAnchor.constraint(equalToConstant: helpBadgeSize.width).isActive = true
+        helpBadge.heightAnchor.constraint(equalToConstant: helpBadgeSize.height).isActive = true
+
+        let spacer = UIBarButtonItem(barButtonSystemItem: .fixedSpace, target: nil, action: nil)
+        spacer.width = helpButtonMarginSpacerWidth
+
+        let barButton = UIBarButtonItem(customView: customView)
+        navigationItem.rightBarButtonItems = [spacer, barButton]
+    }
+
+    // MARK: - UIViewControllerTransitioningDelegate
+
+    /// Displays the support vc.
+    ///
+    func displaySupportViewController(sourceTag: SupportSourceTag) {
+        let controller = SupportViewController()
+        controller.sourceTag = sourceTag
+
+        let navController = UINavigationController(rootViewController: controller)
+        navController.navigationBar.isTranslucent = false
+        navController.modalPresentationStyle = .formSheet
+
+        navigationController?.present(navController, animated: true, completion: nil)
+    }
+}
